requirements.txt
click==6.7
Flask==0.12
itsdangerous==0.24
Jinja2==2.9.4
MarkupSafe==0.23
mwparserfromhell==0.4.4
numpy==1.12.0
olefile==0.44
Pillow==4.0.0
protobuf==3.1.0.post1
psycopg2==2.6.2
requests==2.12.5
scikit-learn==0.18.1
scipy==0.18.1
six==1.10.0
sklearn==0.0
tensorflow==0.12.1
Werkzeug==0.11.15
